add INVOKED
Invoked refers to functions and blocks defined by DEFINE-INVOKED or
FLET-INVOKED. Invoked frames may be recorded in response to
asynchronous events, and at replay the presence of its in-event
triggers the execution of the function associated with the name of
the event.

On one hand, FRAMED, CHECKED, REPLAYED or plain JOURNALED have
@IN-EVENTS that are always predictable from the code and the
preceding events. The control flow - on the level of recorded frames
- is deterministic in this sense. On the other hand, Invoked encodes
in its IN-EVENT what function to call next, introducing
non-deterministic control flow.

By letting events choose the code to run, Invoked resembles typical
[Event Sourcing][event-sourcing] frameworks. When Invoked is used
exclusively, the journal becomes a sequence of events. In contrast,
JOURNALED and its wrappers put code first, and the journal will be a
projection of the call tree.